+++ packages/core-components/src/components/search/search.tsx
@@ -64,7 +64,7 @@ export class SearchComponent {
             slot="end"
             variant="primary"
             onClick={this.onSearchClicked}>
-            <b2b-icon icon="b2b_icon-search"></b2b-icon>
+            <b2b-icon icon="b2b_icon-search" clickable={true}></b2b-icon>
           </b2b-button>
         </b2b-input-group>
       </Host>