+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# [1.4.0-beta.1](https://github.com/otto-de/b2b-design-system/compare/v1.3.3-beta.1...v1.4.0-beta.1) (2023-06-12)
+
+
+### Bug Fixes
+
+* **Icon:** fix icons markup typo ([6b2be76](https://github.com/otto-de/b2b-design-system/commit/6b2be76301caa901575af24735457e6fac0c27b6))
+
+
+### Features
+
+* **Rounded-Icon:** [B2BDS-181] add rounded-icon component ([ad3d7f3](https://github.com/otto-de/b2b-design-system/commit/ad3d7f3a3b43c426aa4fef9552c9f1b2368726a3))
+* **Wizard-Icon:** [B2BDS-181] add wizard-icon component ([da99cd2](https://github.com/otto-de/b2b-design-system/commit/da99cd24e877b2786608586b504c4a929be2bf0c))
+
 ## [1.3.3](https://github.com/otto-de/b2b-design-system/compare/v1.3.2...v1.3.3) (2023-06-08)